Fix #24 - Missing Objects on listing 'dirty' files (submodules)
The diff-representing code attempted to lookup all affected objects
to display their object size, but if the tree entry was a submodule,
would crash as the object id was actually a commit in an external
repo, and consequently 'missing'.

Discovered against the chromium-src repo, which contains a few
submodules...

Also, tweak to use 'view' so that not *all* objects are textually
summarised when just showing the shortened list of the first two
or three.

https://github.com/rtyley/bfg-repo-cleaner/issues/24